tests: added tests for different branches to ensure test coverage
